refactor: console, web UI, common 코드 폴더 분리
