+++ __tests__/validators/RestartValidator.test.js
@@ -1,6 +1,6 @@
-import { RESTART } from '../../src/constants/Configurations';
-import { ERROR_MESSAGE } from '../../src/constants/Messages';
-import RestartValidator from '../../src/validators/RestartValidator';
+import { RESTART } from '../../src/common/constants/Configurations';
+import { ERROR_MESSAGE } from '../../src/common/constants/Messages';
+import RestartValidator from '../../src/common/validators/RestartValidator';
 
 describe('재시작 여부 입력 검증 테스트', () => {
   describe('정상 케이스', () => {
