+++ __tests__/validators/WinningNumbersValidator.test.js
@@ -1,11 +1,11 @@
-import { KEY, LOTTO } from '../../src/constants/Configurations';
-import { ERROR_MESSAGE } from '../../src/constants/Messages';
-import { validateCount } from '../../src/validators/validate';
+import { KEY, LOTTO } from '../../src/common/constants/Configurations';
+import { ERROR_MESSAGE } from '../../src/common/constants/Messages';
+import { validateCount } from '../../src/common/validators/validate';
 import {
   validateTypeAll,
   validateRangeAll,
   WinningNumbersValidator,
-} from '../../src/validators/WinningNumbersValidator';
+} from '../../src/common/validators/WinningNumbersValidator';
 
 describe('당첨 번호 검증', () => {
   describe('정상 케이스', () => {
