+++ src/common/constants/Configurations.js
