+++ src/common/constants/MainConstants.js
