+++ src/common/constants/WinningResultModalConstants.js
