+++ src/common/domains/Lotto.js
