+++ src/common/domains/LottoMachine.js
