+++ src/common/domains/WinningResult.js
