+++ src/step1-index.js
@@ -1,4 +1,4 @@
-import App from './App.js';
+import App from './console/App.js';
 
 const app = new App();
 await app.run();
