+++ src/step2-index.js
@@ -1,5 +1,5 @@
-import Main from './components/Main/Main.js';
-import { SELECTORS } from './constants/WinningResultModalConstants.js';
+import Main from './web/components/Main/Main.js';
+import { SELECTORS } from './common/constants/WinningResultModalConstants.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   new Main(SELECTORS.MAIN).render();
