+++ src/web/components/LottoList/lotto.css
