+++ src/web/components/WinningInputsForm/template.js
