+++ src/web/components/WinningInputsForm/winningInputs.css
