+++ src/web/components/WinningResultModal/modal.css
