+++ src/web/components/WinningResultModal/template.js
