+++ src/web/components/core/ViewComponent.js
